Modified specint dir to work with new test_loader makefiles
